+++ cs/TagsCloudVisualization/CircularCloudLayouter.cs
@@ -7,7 +7,7 @@ public class CircularCloudLayouter
     private readonly List<Rectangle> placedWords = new();
     private readonly Point center;
     private readonly Size resolution;
-    public int CloudArea { get; private set; }
+    public int AllRectanglesArea { get; private set; }
     public double MaxRadius { get; private set; }
 
     public CircularCloudLayouter()
@@ -28,6 +28,11 @@ public class CircularCloudLayouter
         this.center = center;
     }
 
+    public int GetCloudArea()
+    {
+        return placedWords.Sum(rectangle => rectangle.Height * rectangle.Width);
+    }
+
     public Rectangle PutNextRectangle(Size rectangleSize)
     {
         foreach (var coordinate in GetPoints(rectangleSize.Height))
@@ -39,7 +44,7 @@ public class CircularCloudLayouter
                 placedWords.Add(target);
 
                 MaxRadius = Math.Max(MaxRadius, MathWithPoints.DistanceToCenter(coordinate, center));
-                CloudArea += rectangleSize.Height * rectangleSize.Width;
+                AllRectanglesArea += rectangleSize.Height * rectangleSize.Width;
 
                 return target;
             }
